Rename pkg to version
